Add refresh button at the comments.
that button will reload the comments and update it.
change the storing formtat of date and time of commeting, the date
stores as gmt format and shows up as local time format.

move todo to private directory which will not shows up
prevent empty form submiting in html form itself.